fix!: convert tests touching internals to unit tests
This allows removing most of the re-exports in the "test_helpers"
module.

BREAKING CHANGE: this cuts down the test_helpers module, but the
module was never documented to begin with, and it should have been
clear that it was for internal use only.
